# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LMS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: listSubjects
      tags: [subjects]
      summary: List all published subjects
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectsListResponse"
  /subjects/{subjectId}:
    get:
      operationId: getSubject
      tags: [subjects]
      summary: Get subject metadata
      parameters:
        - in: path
          name: subjectId
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subject details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectDetailResponse"
        "404":
          description: Subject not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /subjects/{subjectId}/tree:
    get:
      operationId: getSubjectTree
      tags: [subjects]
      summary: Get subject tree with sections and videos
      parameters:
        - in: path
          name: subjectId
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subject tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectTreeResponse"
        "404":
          description: Subject not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /subjects/{subjectId}/enroll:
    post:
      operationId: enrollSubject
      tags: [subjects]
      summary: Enroll in a subject
      parameters:
        - in: path
          name: subjectId
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "409":
          description: Already enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /subjects/{subjectId}/enrollment:
    get:
      operationId: getEnrollment
      tags: [subjects]
      summary: Check enrollment status
      parameters:
        - in: path
          name: subjectId
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Enrollment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrollmentStatusResponse"
...
  /progress/subjects/{subjectId}:
    get:
      operationId: getSubjectProgress
      tags: [progress]
      summary: Get aggregated subject progress
      parameters:
        - in: path
          name: subjectId
          required: true
          schema:
            type: integer
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subject progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubjectProgressResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    SubjectSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        isPublished:
          type: boolean
        createdAt:
          type: string
          format: date-time
        totalVideos:
          type: integer
        totalSections:
          type: integer
      required:
        - id
        - title
        - slug
        - isPublished
        - createdAt
        - totalVideos
        - totalSections
    SubjectsListResponse:
      type: object
      properties:
        subjects:
          type: array
          items:
            $ref: "#/components/schemas/SubjectSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - subjects
        - total
        - page
        - limit
    SubjectDetailResponse:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        isPublished:
          type: boolean
        createdAt:
          type: string
          format: date-time
        totalVideos:
          type: integer
        totalSections:
          type: integer
        isEnrolled:
          type: boolean
      required:
        - id
        - title
        - slug
        - isPublished
        - createdAt
        - totalVideos
        - totalSections
        - isEnrolled
    TreeVideo:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        orderIndex:
          type: integer
        durationSeconds:
          type: integer
          nullable: true
        locked:
          type: boolean
        isCompleted:
          type: boolean
        lastPositionSeconds:
          type: integer
          nullable: true
      required:
        - id
        - title
        - orderIndex
        - locked
        - isCompleted
    TreeSection:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        orderIndex:
          type: integer
        videos:
          type: array
          items:
            $ref: "#/components/schemas/TreeVideo"
      required:
        - id
        - title
        - orderIndex
        - videos
    SubjectTreeResponse:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/TreeSection"
      required:
        - id
        - title
        - sections
    EnrollmentStatusResponse:
      type: object
      properties:
        isEnrolled:
          type: boolean
        enrolledAt:
          type: string
          format: date-time
          nullable: true
      required:
        - isEnrolled
...
    SubjectProgressResponse:
      type: object
      properties:
        subjectId:
          type: integer
        totalVideos:
          type: integer
        completedVideos:
          type: integer
        percentComplete:
          type: number
        lastVideoId:
          type: integer
          nullable: true
        lastPositionSeconds:
          type: integer
          nullable: true
      required:
        - subjectId
        - totalVideos
        - completedVideos
        - percentComplete